+++ docs/changes.md
@@ -2,6 +2,7 @@
 
 2025/05/29
   * 完善switch控件中max_xoffset_ratio属性文档（感谢俊杰提供补丁)。
+  * 修复text_selector控件设置mask_easing属性不生效的问题（感谢俊杰提供补丁)。
 
 2025/05/28
   * 完善 tk_object_exec_ex(感谢兆坤提供补丁)。